change monster asset load timing
